추가 과제 완료(+구글)

+++ src/components/TodoList_admin.js
@@ -1,7 +1,7 @@
 import React, {useState,useEffect} from "react";
-import {useSession,signOut} from "next-auth/react";
-
-import TodoItem from "./Todoitem"
+import {useRouter} from "next/router";
+import {useSession} from "next-auth/react";
+import TodoItem from "./Todoitem_admin"
 import styles from "../styles/TodoList.module.css";
 
 //firebase
@@ -17,63 +17,38 @@ import {
     orderBy,
     where,
 } from "firebase/firestore";
-import { publicRuntimeConfig } from "../../next.config";
 
 const todoCollection = collection(db,"todos");
 const ids = collection(db,"ids");
 
-const TodoList = () => {
+const TodoList_admin = () => {
     const today=new Date();
     const today_str=`${today.getFullYear()}-${(today.getMonth()+1).toString().padStart(2,'0')}-${today.getDate().toString().padStart(2,'0')}`
     const [itemInput,setItem] = useState("");
     const [dateInput,setDate] = useState(today_str);
     const [todos,setTodos] = useState([]);
-    const [but,setbut]=useState('추가하기');
+    const [but,setbut]=useState('리마인더 추가하기');
     const [modid,setmodid] =useState("");
 
-    const [privat,setprivat]=useState(false);
-    const [baseid,setbaseid]=useState("");
+    const [idlst,setids]=useState([]);
 
-    const {data} =useSession();
+    const router=useRouter();
 
-
-    //userID 있는지 없는지 확인
-    const get_privat=async()=>{
+    const get_ids=async()=>{
         const q2=query(ids);
         const result2=await getDocs(q2);
-        setprivat(false);
+        const new_ids=[];
         result2.docs.forEach((doc)=>{
-            // console.log(doc.data().userId,data?.user?.id,doc.data().userId==data?.user?.id);
-            if (doc.data().userId==data?.user?.id){
-                setbaseid(doc.id)
-                setprivat(true);
-                }
+            console.log(doc.data().userId);
+            new_ids.push(doc.data().userId);
         });
-    }
-
-    const add_id_privat=async()=>{
-        console.log(privat);
-        if (privat){
-            const del_doc=doc(ids,baseid);
-            await deleteDoc(del_doc);
-            alert("Now public");
-            setprivat(false);
-        } else {
-            const docRef = await addDoc(ids, {
-                userId: data?.user?.id,
-            });
-            alert("Now private");
-            setprivat(true);
-        }
+        setids(new_ids);
     }
 
     const getTodos = async () => {
 
-        if (!data?.user?.name) return;
-
         const q=query(
-            todoCollection,
-            where("userId","==",data?.user?.id));
+            todoCollection);
         // const q = query(collection(db, "todos"), where("user", "==", user.uid));
         // const q = query(todoCollection, orderBy("date", "desc"));
 
@@ -91,17 +66,16 @@ const TodoList = () => {
 
     useEffect(()=>{
         getTodos();
-    },[data]);
-    useEffect(()=>{
-        get_privat();
-    },[data]);
+        get_ids();
+        console.log(idlst);
+    },[]);
 
     const addTodo = async () => {
         if (itemInput.trim()==="" || dateInput==="") return;
 
-        if (but==="추가하기"){
+        if (but==="리마인더 추가하기"){
             const docRef = await addDoc(todoCollection, {
-                userId: data?.user?.id,
+                userId:"admin",
                 item:itemInput,
                 date:dateInput,
                 completed:false,
@@ -121,7 +95,7 @@ const TodoList = () => {
             // setTodos(todos.map((todo)=>todo.id==modid? {...todo,item:itemInput,date:dateInput}:todo));
             newTodos.sort((a,b)=>new Date(a.date).getTime()-new Date(b.date).getTime());
             setTodos(newTodos);
-            setItem("");setDate(today_str);setbut("추가하기")
+            setItem("");setDate(today_str);setbut("리마인더 추가하기")
             setmodid("");
         }
         
@@ -130,42 +104,70 @@ const TodoList = () => {
     const toggleTodo = (id) => {
         // setTodos(todos.map((todo)=>todo.id===id? {...todo,completed:!todo.completed}:todo));
         const newTodos=todos.map((todo)=>{
-            if (todo.id==id){
-                const todoDoc=doc(todoCollection,id);
-                updateDoc(todoDoc,{completed:!todo.completed});
-                return {...todo,completed:!todo.completed}
-            } else {
+            if (todo.id==id && idlst.includes(todo.userId)){
+                alert("편집권한이 없습니다!")
+                console.log("!");
                 return todo;
+            } else {
+                if (todo.id==id){
+                    const todoDoc=doc(todoCollection,id);
+                    updateDoc(todoDoc,{completed:!todo.completed});
+                    return {...todo,completed:!todo.completed}
+                } else {
+                    return todo;
+                }
             }
+                
         });
         setTodos(newTodos);
     }
 
     const modiTodo = (id) => {
-        const selected=todos.filter((todo)=>todo.id===id)[0]
-        setItem(selected.item);setDate(selected.date);setbut("수정하기");
-        setmodid(id);
+        let flag=true;
+        todos.forEach((todo)=>{
+            if (todo.id==id && idlst.includes(todo.userId)){
+                alert("편집권한이 없습니다!");
+                flag=false;
+                // console.log("!");
+            }
+        })
+        if (flag){
+            const selected=todos.filter((todo)=>todo.id===id)[0]
+            setItem(selected.item);setDate(selected.date);setbut("리마인더 수정하기");
+            setmodid(id);
+        }
+        
     }
 
     const delTodo = (id) => {
         // console.log(todos);
         // console.log(id);
-        if (but=="수정하기"){
-            const selected=todos.filter((todo)=>todo.id===modid)[0];
-            if (id===selected.id){
-                setItem("");setDate(today_str);setbut("추가하기");
-                setmodid("");
+        let flag=true;
+        todos.forEach((todo)=>{
+            if (todo.id==id && idlst.includes(todo.userId)){
+                alert("편집권한이 없습니다!");
+                flag=false;
+                // console.log("!");
             }
+        })
+        if (flag){
+            if (but=="리마인더 수정하기"){
+                const selected=todos.filter((todo)=>todo.id===modid)[0];
+                if (id===selected.id){
+                    setItem("");setDate(today_str);setbut("리마인더 추가하기");
+                    setmodid("");
+                }
+            }
+            const todoDoc=doc(todoCollection,id);
+            deleteDoc(todoDoc);
+            setTodos(todos.filter((todo)=>todo.id!==id));
         }
-        const todoDoc=doc(todoCollection,id);
-        console.log("!",todoDoc);
-        deleteDoc(todoDoc);
-        setTodos(todos.filter((todo)=>todo.id!==id));
     }
     //렌더링
     return (
-        <div className="max-w-600 w-2/5 mx-auto p-10 bg-white shadow-xl rounded-lg">
-            <h1 className="text-4xl text-orange-500 mb-7 font-extrabold underline underline-light underline-offset-1 decoration-double">{data?.user?.name}'s Reminder</h1>
+        <div className="bg-gray-700 w-2/5 mx-auto p-10 shadow-xl rounded-lg">
+            <h1 className="text-4xl text-black mb-7 font-extrabold underline underline-light underline-offset-1 decoration-double">
+                ADMIN's Reminder</h1>
             <div className={styles.input_group}>
                 <input
                 type="text"
@@ -180,22 +182,16 @@ const TodoList = () => {
                 onChange={(e)=>{setDate(e.target.value)}}
                 />
             </div>
-            <button className={but=="추가하기"? 
-            "shadow-lg w-20 justify-self-end p-1 mb-4 bg-blue-500 text-white border border-blue-500 \
+            <button className={but=="리마인더 추가하기"? 
+            "shadow-lg w-40 justify-self-end p-1 mb-4 bg-blue-500 text-white border border-blue-500 \
             rounded hover:bg-white hover:text-blue-500"
-            :"shadow-lg w-20 justify-self-end p-1 mb-4 bg-green-500 text-white border border-green-500 \
+            :"shadow-lg w-40 justify-self-end p-1 mb-4 bg-green-500 text-white border border-green-500 \
             rounded hover:bg-white hover:text-green-500"
             } onClick={addTodo}>
                 {but}
             </button>
-            <button className={privat? "shadow-lg w-20 justify-self-end p-1 mb-4 ml-5 bg-orange-500 text-white border border-orange-500 \
-            rounded hover:bg-white hover:text-orange-500"
-            : "shadow-lg w-20 justify-self-end p-1 mb-4 ml-5 bg-green-500 text-white border border-green-500 \
-            rounded hover:bg-white hover:text-green-500"
-            } onClick={()=>add_id_privat()}>{privat? "private":"public"}</button>
-            <button className={"shadow-lg w-20 justify-self-end p-1 mb-4 ml-5 bg-red-500 text-white border border-red-500 \
-            rounded hover:bg-white hover:text-red-500"
-            } onClick={()=>signOut()}>Sign out</button>
+            <button className="shadow-lg w-15 justify-self-end p-1 mb-4 ml-5 bg-blue-500 text-white border border-blue-500 \
+            rounded hover:bg-white hover:text-blue-500" onClick={()=>{router.replace("../../auth/signin")}}>Back to Main</button>
             <ul>
                 {todos.map((todo)=>(
                     <TodoItem
@@ -211,4 +207,4 @@ const TodoList = () => {
     );
 };
 
-export default TodoList;
+export default TodoList_admin;
